fixes spurious test error in PR #796
server.js was throwing an error, but only when run *after* the
net_utils tests. I tracked it down to this line causing the problem:
```js
config.get('smtp.ini').main.public_ip
```

This ends up being an issue related to config caching. `net_utils`
reads a copy of smtp.ini to see if a public IP is defined there but it
does’t pass in the same options as the server.js file. Later, when
server.js reads that same file, with options, it gets the cached result
without them.